Move http2 files around
