+++ http2/src/main/java/com/blazemeter/jmeter/http2/sampler/HTTP2SampleResult.java
@@ -16,7 +16,7 @@
  *
  */
 
-package blazemeter.jmeter.plugins.http2.sampler;
+package com.blazemeter.jmeter.http2.sampler;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
