+++ http2/src/main/java/com/blazemeter/jmeter/http2/visualizers/RenderAsJSON.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package blazemeter.jmeter.plugins.http2.visualizers;
+package com.blazemeter.jmeter.http2.visualizers;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
