+++ http2/src/main/java/com/blazemeter/jmeter/http2/visualizers/RenderAsRegexp.java
@@ -20,7 +20,7 @@
 /**
  *
  */
-package blazemeter.jmeter.plugins.http2.visualizers;
+package com.blazemeter.jmeter.http2.visualizers;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
