+++ http2/src/main/java/com/blazemeter/jmeter/http2/visualizers/RenderAsXML.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package blazemeter.jmeter.plugins.http2.visualizers;
+package com.blazemeter.jmeter.http2.visualizers;
 
 import java.awt.Component;
 import java.awt.GridLayout;
